gin-grpc implemented

+++ src/Structs/Structs.go
@@ -3,7 +3,7 @@ package Structs
 
 type Item struct {
 	Name     string  `json:"Name"`
-	Price    float64 `json:"Price"`
+	Price    float32 `json:"Price"`
 	Quantity int64   `json:"Quantity"`
 }
 
@@ -12,7 +12,7 @@ type Order struct {
 	CustomerID int64 `json:"CustomerID"`
 	Restaurant string `json:"Restaurant"`
 	ItemLine   []Item `json:"ItemLine"`
-	Price      float64 `json:"Price"`
+	Price      float32 `json:"Price"`
 	Quantity   int64 `json:"Quantity"`
 	Discount   int64 `json:"Discount"`
 	Date       string `json:"Date"`
